servico-pedidos/requirements.txt
flask==2.3.3
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0